requirements.txt
numpy==1.24.1 # used for the actual model code/weights
tensorflow==2.11.0; sys_platform != 'darwin' or platform_machine != 'arm64' # used to load the gpt-2 weights from the open-ai tf checkpoint
tensorflow-macos==2.11.0; sys_platform == 'darwin' and platform_machine == 'arm64' # used on M1 Macbooks
regex==2017.4.5 # used by the bpe tokenizer
gpt3_tokenizer==0.1.5 # used for tokenization
tokenizers==0.15.2 # used for tokenization
datasets==2.18.0 # used for data loading and storage
pyarrow==15.0.2 # used for data loading and storage
pyarrow-hotfix==0.6 # used for data loading and storage
requests==2.27.1 # used to download gpt-2 files from openai
tqdm==4.64.0 # progress bar to keep your sanity
fire==0.5.0 # easy CLI creation
aiohttp==3.9.3 # used for async HTTP requests
aiosignal==1.3.1 # used for async signal handling
anyio==4.3.0 # used for async I/O operations
transformers==4.40.0.dev0 # used for transformer-based models
torch==2.0.1+cu117 # used for PyTorch-based models
torchaudio==2.0.2+cu117 # used for audio processing
torchvision==0.15.2+cu117 # used for computer vision tasks
scikit-learn==1.4.1.post1 # used for machine learning tasks
seaborn==0.13.2 # used for data visualization
accelerate==0.29.1 # used for accelerating computations
appdirs==1.4.4 # used for determining application directories
asttokens==2.4.1 # used for parsing Python code
async-timeout==4.0.3 # used for async timeout handling
attrs==23.2.0 # used for attribute-based programming
bitsandbytes==0.43.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numba==0.59.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
parso==0.8.4
pillow==10.3.0
pip==22.2.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.1
responses==0.18.0
rsa==4.9
safetensors==0.4.2
setproctitle==1.3.3
setuptools==63.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tiktoken==0.7.0
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
utils==1.0.2
wandb==0.16.6
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4